Create header section components and add styling.
Created "MainHeader" client-component and its associated module css
styles. This component is conformed by a title+logo section and a
navigation bar.

A react-icon has been used as the logo. To that end, the "react-icons"
library has been installed as a dependency.

The navigation bar is conformed by a list of "NavItem" sub-components,
which make the code more simple and modular.

General level styling has been added to the header section components,
but some features are pending.