Style "blog" page and add "PostCard" component.
Blog page still fetches the posts from the API but now they are derived
and rendered using the "PostCard" server component, which improves the
SEO of the page.

CSS styling has been added to the "PostCard" component and the "blog"
page. Posts are shown in a grid layout with a card for each post.

The "PostCard" creates a slug for each post by taking the title of the
post and replacing blank-space separators by dashes. The slug is used to
create a link to the post page.

New posts have been added to the "posts" directory. The posts metadata
images are fetched from the Unsplash website, which was authorized
within the "next.config.mjs" file.