Add filtering logic for blog page component.
The blog posts section has 4 different filters. 3 dropdown filters
related to date, author, and categoy, and one search bar which can be
used by the user to filter by any words found within the posts metadata.

Two different states are used for the blog page component. The first one
is an object containing keys for all posts, and the posts filtered by
date, author and category. The second one is a state that receives the
current value of the search camp, allowing to filter accordingly.

Having a different piece of state for each of the dropdown filters
ensures the posts value for each filter is independent from the rest,
allowing to use a helper function (getCommonFilteredPosts) to create an
array with the posts that fit all filters requirements. This array is
then passed through the search bar filter (if needed), and then finally
used to render the required posts (but only if the user has filtered
through the dropdowns, otherwise the allPosts state is used directly).